moved contacts state from Table component to App component
